requirements.txt
pygithub
load_dotenv